Finish draft of DOVIEWCHANGE handler
Finish a rough draft of the handler for DOVIEWCHANGE on the primary,
up until we send the STARTVIEW message to all new backups.

Start thinking about where to clear the relevant metadata for each
stage of the view change protocol, i.e. where to reset the counters
and caches we are using to determine when quorum is reached and what
messages to ignore.