Start adding notes on VR module design
We intend to write a simple distributed key-value store service using
our VR implementation. To do this, we would prefer to organize the VR
module such that it is agnostic to the operation execution/state
change specifics of the service it is being used to replicate. This
will require some possible re-thinking and re-organization of the VR
module code.